+++ NEWS.md
@@ -1,7 +1,9 @@
 # rchallenge 1.2.0.9000
 - `output_dir` argument of `publish` function now defaults to `"index.html"`. Useful for hosting the challenge on a GitHub repo with Github pages.
-- glyphicon is deprecated. use fontawsome instead of glyphicon.
-- print_readerr displays a table.
+- `glyphicon` is deprecated. use fontawsome instead of glyphicon.
+- `print_readerr` displays a table.
+- `get_best` returns a single data.frame instead of a list with one data.frame per metric. the ranking can be based on several metrics in a specific order to break ties.
+- `update_rank_diff` and `print_leaderboard` take a single data.frame as input
 
 # rchallenge 1.2.0 (05-10-2016)
 - `output_dir` argument of `publish` function now defaults to the `input` directory instead of `"~/Dropbox/Public"` because Dropbox rendering of HTML content is discontinued.